src/main/java/com/SpringBoot/web/HelloController.java
package com.SpringBoot.web;

import com.SpringBoot.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class HelloController {
    @GetMapping("/hello") //get요청을 받음. 함수 앞에 @어노테이션을 사용하면 되는 듯. Spring에서는 어노테이션이 많은 기능을 하는 것 같다.
    public String hello() {
        return "hello";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name,
                                     @RequestParam("amount") int amount) {
        return new HelloResponseDto(name, amount);
    }
}
